requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
python-dotenv==1.0.1
Werkzeug==2.3.7
